Coloquei em componentes detalhes e gestao de cliente
